remove explicit bits

+++ resources.go
@@ -40,7 +40,7 @@ type Resource struct {
 	CreatedAt       string   `json:"created"`
 	UpdatedAt       string   `json:"updated"`
 	Name            string   `json:"name"`
-	ClientID        int64    `json:"client"`
+	ClientID        int      `json:"client"`
 	Active          bool     `json:"active"`
 	Enabled         bool     `json:"enabled"`
 	WAF             bool     `json:"is_waf_enabled"`
@@ -55,10 +55,10 @@ type Resource struct {
 	MultipleOrigins bool     `json:"feature_multiple_origins"`
 	WidlcardAliases bool     `json:"feature_wildcard_aliases"`
 	SSLType         string   `json:"ssl_type"`
-	SSLExpire       int64    `json:"service_ssl_expire"`
+	SSLExpire       int      `json:"service_ssl_expire"`
 	SSLStatus       string   `json:"service_ssl_status"`
 	TLSEnabled      []string `json:"tls_enabled"`
-	WaitForLE       int64    `json:"wait_for_le"`
+	WaitForLE       int      `json:"wait_for_le"`
 }
 
 // ResourceCreateRequest represents a request to create a DDoS protection resource
@@ -100,10 +100,10 @@ type ResourceUpdateRequest struct {
 
 // ResourceListOptions specifies the optional query parameters to List method
 type ResourceListOptions struct {
-	Limit           int32  `url:"limit,omitempty" validate:"omitempty"`
-	Offset          int32  `url:"offset,omitempty" validate:"omitempty"`
+	Limit           int    `url:"limit,omitempty" validate:"omitempty"`
+	Offset          int    `url:"offset,omitempty" validate:"omitempty"`
 	Ordering        string `url:"ordering,omitempty" validate:"omitempty"`
-	ClientID        int64  `url"client,omitempty" validate:"omitempty"`
+	ClientID        int    `url"client,omitempty" validate:"omitempty"`
 	Name            string `url:"name,omitempty" validate:"omitempty"`
 	Active          bool   `url:"active,omitempty" validate:"omitempty"`
 	MultipleOrigin  bool   `url:"feature_multiple_origins,omitempty" validate:"omitempty"`
